Add support for authors (outside of data)
Allow the creation of a new category: the authors.

The change patches the base theme to allow the usage of the author
sidebar which comes from minimal-mistakes.
That sidebar already supports authors added through the data files,
but this new authors type will be managed through separate files
instead (thanks to Netlify CMS).